moved distributor statement summary to dedicated component
