+++ apps/waterfall/waterfall/src/app/dashboard/statement/view/view.module.ts
@@ -2,45 +2,33 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { ReactiveFormsModule } from '@angular/forms';
 
 // Blockframes
 import { StatementHeaderModule } from '@blockframes/waterfall/components/statement-header/statement-header.module';
-import { PricePerCurrencyModule } from '@blockframes/ui/price-per-currency/price-per-currency.module';
-import { TableModule } from '@blockframes/ui/list/table/table.module';
+import { StatementDistributorSummaryModule } from '@blockframes/waterfall/components/statement-distributor/summary/summary.module';
 
 // Components
 import { StatementViewComponent } from './view.component';
 
 // Material
-import { MatDividerModule } from '@angular/material/divider';
-import { MatExpansionModule } from '@angular/material/expansion';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
 import { MatTooltipModule } from '@angular/material/tooltip';
 
 @NgModule({
   declarations: [StatementViewComponent],
   imports: [
     CommonModule,
-    ReactiveFormsModule,
 
     // Blockframes
     StatementHeaderModule,
-    PricePerCurrencyModule,
-    TableModule,
+    StatementDistributorSummaryModule,
 
     // Material
-    MatDividerModule,
-    MatExpansionModule,
     MatIconModule,
     MatButtonModule,
     MatSnackBarModule,
-    MatFormFieldModule,
-    MatInputModule,
     MatTooltipModule,
 
     // Routing
